requirements.txt
fastapi 
uvicorn[standard] 
pillow 
numpy 
opencv-python 
pydantic